security-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
cryptography==42.0.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
